+++ spec/directions/move_solutions_to_answer_key/answer_key_inner_container_spec.rb
@@ -1,8 +1,8 @@
-# frozen-string-literal: true
+# frozen_string_literal: true
 
 require 'spec_helper'
 
-RSpec.describe Kitchen::Directions::BookAnswerKeyContainer do
+RSpec.describe Kitchen::Directions::AnswerKeyInnerContainer do
   let(:book) do
     book_containing(html:
       <<~HTML
@@ -16,15 +16,25 @@ RSpec.describe Kitchen::Directions::BookAnswerKeyContainer do
     )
   end
 
-  it 'v1 works' do
-    expect(described_class.v1(book: book)).to match_normalized_html(
+  let(:append_to) do
+    new_element(
       <<~HTML
-        <div class="os-eob os-solutions-container" data-type="composite-chapter" data-uuid-key=".solutions">
-          <h1 data-type="document-title">
-            <span class="os-text">Answer Key</span>
-          </h1>
+        <div class="heyhey"></div>
+      HTML
+    )
+  end
+
+  it 'v1 works for solution (plural)' do
+    expect(
+      described_class.v1(chapter: book.chapters.first, metadata_source: metadata_element, append_to: append_to)
+    ).to match_normalized_html(
+      <<~HTML
+        <div class="os-eob os-solutions-container" data-type="composite-page" data-uuid-key=".solutions1">
+          <h2 data-type="document-title">
+            <span class="os-text">Chapter 1</span>
+          </h2>
           <div data-type="metadata" style="display: none;">
-            <h1 data-type="document-title" itemprop="name">Answer Key</h1>
+            <h1 data-type="document-title" itemprop="name">Chapter 1</h1>
             <span data-type="revised" id="revised_copy_1">Revised</span>
             <span data-type="slug" id="slug_copy_1">Slug</span>
             <div class="authors" id="authors_copy_1">Authors</div>
@@ -38,15 +48,19 @@ RSpec.describe Kitchen::Directions::BookAnswerKeyContainer do
     )
   end
 
-  it 'v1 generates container with solution (singular) class' do
-    expect(described_class.v1(book: book, solutions_plural: false)).to match_normalized_html(
+  it 'v1 works for solution (singular)' do
+    expect(
+      described_class.v1(
+        chapter: book.chapters.first, metadata_source: metadata_element, append_to: append_to, solutions_plural: false
+      )
+    ).to match_normalized_html(
       <<~HTML
-        <div class="os-eob os-solution-container" data-type="composite-chapter" data-uuid-key=".solution">
-          <h1 data-type="document-title">
-            <span class="os-text">Answer Key</span>
-          </h1>
+        <div class="os-eob os-solution-container" data-type="composite-page" data-uuid-key=".solution1">
+          <h2 data-type="document-title">
+            <span class="os-text">Chapter 1</span>
+          </h2>
           <div data-type="metadata" style="display: none;">
-            <h1 data-type="document-title" itemprop="name">Answer Key</h1>
+            <h1 data-type="document-title" itemprop="name">Chapter 1</h1>
             <span data-type="revised" id="revised_copy_1">Revised</span>
             <span data-type="slug" id="slug_copy_1">Slug</span>
             <div class="authors" id="authors_copy_1">Authors</div>