+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Add `BakeAllNumberedExerciseTypes` direction for easier baking of compound sections (minor)
 * Add `solution_stays_put` option for `BakeNumberedExercise` (minor)
+* Refactor: moves all `Answer key strategies` that are book-specific to the
+recipes side and keeps the `Default Strategy`in kitchen (major)
 
 ## [14.0.0] - 2021-10-22